chore: refactoring jsii2schema
